coaches/requirements.txt
playwright
pyyaml
requests
beautifulsoup4
lxml